+++ src/components/campaignSlider/CampaignSlider.jsx
@@ -8,140 +8,123 @@ import "swiper/css/autoplay";
 const slides = [
   {
     image:
-      "https://www.dominos.de/ManagedAssets/DE/product/PCSP/DE_PCSP_all_hero_11910.jpg?v-910499353",
-    title: "This is title 1",
+      "https://www.birazbira.com/wp-content/uploads/2019/05/bira-nedir-1.jpg",
+    title: "Bira",
   },
   {
     image:
-      "https://www.dominos.de/ManagedAssets/DE/product/PSSA/DE_PSSA_all_hero_10571.jpg?v-377603643",
-    title: "This is title 2",
+      "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTLKavFCvqw_lsk4i71FLZCIsPCm4nOwPGPHbrBIAiYw1hI9g9PPLYorlX4IErtF--64nE&usqp=CAU",
+    title: "Dürüm",
   },
   {
     image:
-      "https://www.dominos.de/ManagedAssets/DE/product/PDUT/DE_PDUT_all_hero_11734.png?v-124775118",
-    title: "This is title 3",
+      "https://cdn.yemek.com/mnresize/1250/833/uploads/2020/11/firin-sutlac-son-tarifi.jpg",
+    title: "Sütlaç",
   },
   {
     image:
-      "https://www.dominos.de/ManagedAssets/DE/product/PBOS/DE_PBOS_all_hero_11872.png?v-2113143995",
-    title: "This is title 4",
+      "https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQ9jgLewdqw8eBLeJJ8DKxWSEdDtReGC3Q9VQ&usqp=CAU",
+    title: "Karışık Baklava",
   },
   {
     image:
-      "https://www.dominos.de/ManagedAssets/DE/product/PPON/DE_PPON_all_hero_11872.png?v1872079273",
-    title: "This is title 5",
+      "https://www.unileverfoodsolutions.com.tr/dam/global-ufs/mcos/TURKEY/calcmenu/recipes/TR-recipes/2023/updates/header/AdanaKebap%20-%201260x839.jpg",
+    title: "Kebap",
+  },
+  {
+    image: "https://cdn.getiryemek.com/restaurants/1598533454185_1125x522.jpeg",
+    title: "İskender",
   },
   {
     image:
-      "https://www.dominos.de/ManagedAssets/DE/product/PMEA/DE_PMEA_all_hero_11910.jpg?v1167088276",
-    title: "This is title 6",
-  },
-  {
-    image:
-      "https://www.dominos.de/ManagedAssets/DE/product/PCON/DE_PCON_all_hero_11734.png?v1493403577",
-    title: "This is title 7",
+      "https://etibolmanti.com.tr/wp-content/uploads/2019/05/ETIBOL_MANTI_0056s.jpg",
+    title: "Mantı",
   },
   {
     image:
       "https://www.dominos.de/ManagedAssets/DE/product/EPCUR/DE_EPCUR_all_hero_9880.jpg?v752410959",
-    title: "This is title 8",
+    title: "Makarna",
   },
   {
     image:
-      "https://www.dominos.de/ManagedAssets/DE/product/EPVEG/DE_EPVEG_all_hero_11910.jpg?v789189983",
-    title: "This is title 9",
+      "https://d2lswn7b0fl4u2.cloudfront.net/photos/pg-mushroom-pasta-1678049233.jpg",
+    title: "Makarna",
   },
   {
     image:
-      "https://www.dominos.de/ManagedAssets/DE/product/EPORL/DE_EPORL_all_hero_9880.jpg?v916554548",
-    title: "This is title 10",
+      "https://iasbh.tmgrup.com.tr/9473f4/821/464/0/0/0/0?u=https://isbh.tmgrup.com.tr/sbh/2022/08/17/1660716079480.jpg",
+    title: "Makarna",
+  },
+  {
+    image:
+      "https://iasbh.tmgrup.com.tr/9473f4/821/464/0/0/0/0?u=https://isbh.tmgrup.com.tr/sbh/2022/08/17/1660716079480.jpg",
+    title: "Makarna",
+  },
+  {
+    image:
+      "https://im.haberturk.com/2022/10/23/3531958_4897f18fced57500715225fa9e9ce14e_640x640.jpg",
+    title: "Çay",
   },
 ];
 
 const CampaignSlider = () => {
   return (
-    <div className="w-[60%] mx-auto my-5 ">
-      <h1 className="text-center mb-5">Haftanın Enleri</h1>
-      <Swiper
-        effect="coverflow"
-        spaceBetween={50}
-        grabCursor={true}
-        centeredSlides={true}
-        loop={true}
-        autoplay={{
-          delay: 2500,
-          disableOnInteraction: false,
-        }}
-        slidesPerView={3}
-        coverflowEffect={{
-          rotate: 10,
-          stretch: 0,
-          depth: 100,
-          modifier: 3,
-          slideShadows: true,
-        }}
-        navigation={{
-          nextEl: ".swiper-button-next",
-          prevEl: ".swiper-button-prev",
-          clickable: true,
-        }}
-        pagination={{ el: ".swiper-pagination", clickable: true }}
-        modules={[EffectCoverflow, Pagination, Navigation, Autoplay]}
-        className="swiper_containe"
-      >
-        {slides.map((slide, index) => (
-          <SwiperSlide  key={index}>
-            <img src={slide.image} alt={slide.title} />
-          </SwiperSlide>
-        ))}
-        <div >
-          <div className="swiper-button-prev slider-arrow ">
-            <ion-icon name="arrow-back-outline"></ion-icon>
+    <div className="w-full   my-5 ">
+      <h1 className=" mb-5   text-center  font-semibold text-2xl leading-10 ">
+        Haftanın Enleri
+      </h1>
+      <div className="mx-5 xl:w-[65%] xl:mx-auto">
+        <Swiper
+          effect="coverflow"
+          spaceBetween={50}
+          grabCursor={true}
+          centeredSlides={true}
+          loop={true}
+          autoplay={{
+            delay: 2500,
+            disableOnInteraction: false,
+          }}
+          slidesPerView={3}
+          coverflowEffect={{
+            rotate: 10,
+            stretch: 0,
+            depth: 100,
+            modifier: 3,
+            slideShadows: false,
+          }}
+          navigation={{
+            nextEl: ".swiper-button-next",
+            prevEl: ".swiper-button-prev",
+            clickable: true,
+          }}
+          pagination={{ el: ".swiper-pagination", clickable: true }}
+          modules={[EffectCoverflow, Pagination, Navigation, Autoplay]}
+          className=" rounded-lg "
+        >
+          {slides.map((slide, index) => (
+            <SwiperSlide className="w-[350px] rounded-lg" key={index}>
+            <h1 className=" text-xl text-center font-extrabold  ">
+                {slide.title}
+              </h1>
+               <img
+                className="w-full h-[350px] rounded-lg bg-cover bg-no-repeat  "
+                src={slide.image}
+              />            
+            </SwiperSlide>
+          ))}
+          <div>
+            <div className="swiper-button-prev">
+              <ion-icon name="arrow-back-outline"></ion-icon>
+            </div>
+            <div className="swiper-button-next">
+              <ion-icon name="arrow-forward-outline"></ion-icon>
+            </div>
+            <div className="swiper-pagination"></div>
           </div>
-          <div className="swiper-button-next slider-arrow">
-            <ion-icon name="arrow-forward-outline"></ion-icon>
-          </div>
-          <div className="swiper-pagination"></div>
-        </div>
-      </Swiper>
+        </Swiper>
+      </div>
     </div>
   );
 };
 
 export default CampaignSlider;
-
-// const sliderLeft = () => {
-//   var slider = document.querySelector(".slider");
-//   slider.scrollLeft = slider.scrollLeft - 300;
-// };
-// const sliderRight = () => {
-//   var slider = document.querySelector(".slider");
-//   slider.scrollLeft = slider.scrollLeft + 300;
-// };
-// // useEffect hook'unu kullanarak bir zamanlayıcı oluşturuyoruz
-// useEffect(() => {
-//   let isReversed = false;
-//   // setInterval fonksiyonunu kullanarak 5 saniyede bir sliderRight işlevini çağırıyoruz
-//   const interval = setInterval(() => {
-//     if (sliderRight && !isReversed) {
-//       sliderRight();
-//       if (slider.scrollLeft >= slider.scrollWidth - slider.clientWidth) {
-//         // Sağa doğru sona gelindiğinde geri yönde kaydırmak için durumu değiştir
-//         isReversed = true;
-//       }
-//     } else {
-//       sliderLeft();
-//       if (slider.scrollLeft <= 0) {
-//         // Sola doğru sona gelindiğinde tekrar ileri yönde kaydırmak için durumu değiştir
-//         isReversed = false;
-//       }
-//     }
-//   }, 5000);
-
-//   // useEffect hook'unun geri dönüş fonksiyonu aracılığıyla zamanlayıcıyı temizliyoruz
-//   return () => {
-//     clearInterval(interval);
-//   };
-// }, []);
-
-// style={{ backgroundImage: `url(${slide.image})` }}
