+++ src/index.css
@@ -260,8 +260,6 @@ input {
   background-color: white !important;
   color: #e12a32 !important;
   border: solid 2px black;
-
-
 }
 
 .md-position{
@@ -274,51 +272,6 @@ input {
 
 
 
-/* swiper */
-
-/* .container {
-  max-width: 124rem;
-  padding: 4rem 1rem;
-  margin: 0 auto;
-}
-
-.heading {
-  padding: 1rem 0;
-  font-size: 3.5rem;
-  text-align: center;
-}
-
-.swiper_container {
-  height: 52rem;
-  padding: 2rem 0;
-  position: relative;
-}
-
-.swiper-slide {
-  width: 37rem;
-  height: 42rem;
-  position: relative;
-}
-
-.swiper-slide img {
-  width: 37rem;
-  height: 42rem;
-  border-radius: 2rem;
-  object-fit: cover;
-}
-
-.swiper-slide-shadow-left,
-.swiper-slide-shadow-right {
-  display: none;
-}
-
-.slider-controler {
-  position: relative;
-  bottom: 2rem;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-} */
